requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
install==1.3.5
Pillow==9.5.0
pycodestyle==2.10.0
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
